Fix some bugs in the lookup tables.
Never got a single bug report about mystic draw in all the years. That's why I never fixed anything.
Now with translating the old code to rust I spot tons of issues.

The 1.6 branch should not be used anymore - it's only there for historical reasons.

Rust is the future (even for programs of yesterday).