Implemented extendet 512 font support.
As far as I know there are no other drawing tools supporting that.
Unfortunately the viewing tools miss that as well.

According to http://wiki.synchro.net/ref:xbin
There is only one old MS-DOS program capable of showing files with 512 chars. I tested that in DOSBox and it works there.
So that's a feature for all users who want to produce content for SIMPLEXB.EXE :). Enjoy